pyproject.toml
[project]
name = "delta-exchange"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "websocket-client>=1.8.0",
]
